Split into separate files, add builder, make impl immutable
